[rust] Separate modules into individual crates
