+++ crates/petname/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "multipass-petname"
+version = "0.1.0"
+edition = "2021"
+resolver = "2"
+
+[lib]
+path = "src/lib.rs"
+crate-type = ["staticlib"]
+
+[dependencies]
+cxx = { workspace = true }
+rand = { workspace = true }
+
+[build-dependencies]
+cxx-build = { workspace = true }
