+++ crates/petname/build.rs
@@ -1,8 +1,7 @@
 fn main() {
     cxx_build::bridge("src/lib.rs")
         .flag_if_supported("-std=c++17")
-        .compile("multipass");
+        .compile("multipass-petname");
 
     println!("cargo:rerun-if-changed=src/lib.rs");
-    println!("cargo:rerun-if-changed=src/petname/mod.rs");
 }
