+++ src/daemon/daemon.cpp
@@ -22,6 +22,7 @@
 #include "rust/cxx.h"
 #include "snapshot_settings_handler.h"
 
+#include <multipass-petname/src/lib.rs.h>
 #include <multipass/alias_definition.h>
 #include <multipass/cloud_init_iso.h>
 #include <multipass/constants.h>
@@ -45,7 +46,6 @@
 #include <multipass/settings/bool_setting_spec.h>
 #include <multipass/settings/settings.h>
 #include <multipass/snapshot.h>
-#include <multipass/src/lib.rs.h>
 #include <multipass/ssh/ssh_session.h>
 #include <multipass/sshfs_mount/sshfs_mount_handler.h>
 #include <multipass/top_catch_all.h>
@@ -218,16 +218,14 @@ auto name_from(const std::string& requested_name,
     else
     {
         // Create a Rust petname generator with 2 words and "-" separator
-        auto petname_generator = multipass::new_petname(2, "-");
-        auto rust_name = multipass::make_name(*petname_generator);
-        std::string name{rust_name.c_str(), rust_name.size()};
+        auto petname_generator = multipass::petname::new_petname(2, "-");
+        std::string name = std::string(multipass::petname::make_name(*petname_generator));
         constexpr int num_retries = 100;
         for (int i = 0; i < num_retries; i++)
         {
             if (currently_used_names.find(name) != currently_used_names.end())
             {
-                rust_name = multipass::make_name(*petname_generator);
-                name = std::string{rust_name.c_str(), rust_name.size()};
+                name = std::string(multipass::petname::make_name(*petname_generator));
                 continue;
             }
             return name;
